+++ lib/mainAppBar.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class MainAppBar extends StatelessWidget implements PreferredSizeWidget {
+  final String title;
+  final Color backgroundColor;
+  const MainAppBar({
+    Key? key,
+    required this.title,
+    required this.backgroundColor,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      title: Text(title),
+      titleTextStyle: TextStyle(color: Colors.black, fontSize: 30),
+      centerTitle: true,
+      backgroundColor: backgroundColor,
+      elevation: 0,
+      actions: [
+        IconButton(
+          icon: Icon(Icons.more_vert, color: Colors.black, size: 28),
+          onPressed: () {},
+        )
+      ],
+    );
+  }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(kToolbarHeight);
+}
